+++ src/main/java/com/whiteday/aiecolink/domain/member/model/entity/User.java
@@ -1,5 +1,6 @@
-package com.whiteday.aiecolink.member;
+package com.whiteday.aiecolink.domain.member.model.entity;
 
+import com.whiteday.aiecolink.domain.member.model.Role;
 import jakarta.persistence.*;
 import lombok.*;
 import java.time.LocalDateTime;
@@ -22,7 +23,7 @@ public class User {
     @Column(nullable = false, unique = true, length = 100)
     private String email;
 
-    @Column(nullable = false, length = 50)
+    @Column(nullable = false, length = 255)
     private String password;
 
     @Column(nullable = false)
@@ -37,4 +38,10 @@ public class User {
 
     @Column(nullable = false)
     private LocalDateTime updatedAt;
+
+    public void updateNameAndPassword(String name, String encodedPassword) {
+        this.name = name;
+        this.password = encodedPassword;
+        this.updatedAt = LocalDateTime.now();
+    }
 }
