+++ src/main/java/com/whiteday/aiecolink/domain/member/service/UserService.java
@@ -1,10 +1,16 @@
-package com.whiteday.aiecolink.member;
+package com.whiteday.aiecolink.domain.member.service;
 
+import com.whiteday.aiecolink.domain.member.dto.UserProfileUpdateRequestDto;
+import com.whiteday.aiecolink.domain.member.model.Role;
+import com.whiteday.aiecolink.domain.member.model.entity.User;
+import com.whiteday.aiecolink.domain.member.repository.UserRepository;
 import com.whiteday.aiecolink.global.error.CustomException;
 import com.whiteday.aiecolink.global.error.ErrorCode;
-import com.whiteday.aiecolink.member.dto.UserRequestDto;
-import com.whiteday.aiecolink.member.dto.UserResponseDto;
+import com.whiteday.aiecolink.domain.member.dto.UserProfileResponseDto;
+import com.whiteday.aiecolink.domain.member.dto.UserRequestDto;
+import com.whiteday.aiecolink.domain.member.dto.UserResponseDto;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
@@ -14,6 +20,7 @@ import java.time.LocalDateTime;
 public class UserService {
 
     private final UserRepository userRepository;
+    private final PasswordEncoder passwordEncoder;
 
     public Long registerUser(UserRequestDto requestDto) {
         if (userRepository.existsByEmail(requestDto.getEmail())) {
@@ -45,4 +52,23 @@ public class UserService {
                 .updatedAt(user.getUpdatedAt())
                 .build();
     }
+
+    public void updateUserProfile(User user, UserProfileUpdateRequestDto request) {
+        // 현재 비밀번호 확인
+        if (!passwordEncoder.matches(request.getCurrentPassword(), user.getPassword())) {
+            throw new CustomException(ErrorCode.PASSWORD_MISMATCH);
+        }
+
+        // 이름과 비밀번호 업데이트
+        user.updateNameAndPassword(
+                request.getNewName(),
+                passwordEncoder.encode(request.getNewPassword())
+        );
+
+        userRepository.save(user);
+    }
+    public UserProfileResponseDto getProfile(User user) {
+        return new UserProfileResponseDto(user.getName());
+    }
+
 }
